refactor(requestlog): move statistics queries into store_stats.go
